Adds test case
This commit begins the refactorization of the current dirty
encryptAndSend() function. This commit is the base for fixing issue #51
and an undocumented issue which doesn't send signed emails unless
this.passphrase !== "", and then signs the message twice.
Refactorization will also help solve #46 by getting messy proof of
concepts out of the way.