Fixes coding standard
Resets previous unwanted coding standard changes.

ATTENTION! Any future commit that includes modifications made to the coding
standard that have not been discussed with me will be IGNORED ENTIRELY.
If you want changes made then create an issue in the issue tracker for
discussion BEFORE requesting a pull. I am not going to spend another
minute sorting out bad and uninteresting commits that do nothing but
mess up the code for those that are actually working on it. If you want
to help then grab a real issue. Thank you for understanding.